beaucoup de changements ... faire gaffe

+++ Projet/client/Point.java
@@ -1,16 +1,16 @@
 public class Point{
   //position
-  private float x;
-  private float y;
+  private double x;
+  private double y;
 
-  public Point(float x,float y){
+  public Point(double x,double y){
     this.x = x;
     this.y = y;
   }
 //GETTERS
-  public float getX(){return x;}
-  public float getY(){return y;}
+  public double getX(){return x;}
+  public double getY(){return y;}
   //SETTERS
-  public void setX(float x){this.x = x;}
-  public void setY(float y){this.y = y;}
+  public void setX(double x){this.x = x;}
+  public void setY(double y){this.y = y;}
 }
